requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.1
django-axes==7.0.1
et-xmlfile==1.1.0
gunicorn==23.0.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.8.2
python-dotenv==1.0.1
qrcode[pil]
jpholiday
line-bot-sdk==3.14.0
